Change Tokens update to 2.2
oops!
