--- aspnetcore/fundamentals/change-tokens/sample/ChangeTokenSample.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="poem.txt" CopyToPublishDirectory="PreserveNewest" />
-  </ItemGroup>
-
-</Project>
