--- aspnetcore/fundamentals/change-tokens/sample/Pages/Error.cshtml.cs
@@ -1,19 +1,0 @@
-using System.Diagnostics;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace ChangeTokenSample.Pages
-{
-    public class ErrorModel : PageModel
-    {
-        public string RequestId { get; set; }
-        
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-
-        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public void OnGet()
-        {
-            RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier;
-        }
-    }
-}
