--- aspnetcore/fundamentals/change-tokens/sample/wwwroot/js/site.js
@@ -1,1 +1,0 @@
-﻿// Write your Javascript code.
