+++ aspnetcore/fundamentals/change-tokens/samples/2.x/SampleApp/Pages/_ViewImports.cshtml
