+++ aspnetcore/fundamentals/change-tokens/samples/2.x/SampleApp/appsettings.Development.json
