+++ aspnetcore/fundamentals/change-tokens/samples/2.x/SampleApp/wwwroot/images/console-config-change.png
